+++ include/plateau/granularity_convert/filter_by_city_obj_index.h
@@ -1,8 +1,11 @@
-//
-// Created by Linoal on 2023/10/23.
-//
+#pragma once
+#include <plateau/polygon_mesh/mesh.h>
+#include <libplateau_api.h>
 
-#ifndef LIBPLATEAU_FILTER_BY_CITY_OBJ_INDEX_H
-#define LIBPLATEAU_FILTER_BY_CITY_OBJ_INDEX_H
-
-#endif //LIBPLATEAU_FILTER_BY_CITY_OBJ_INDEX_H
+namespace plateau::granularityConvert {
+    class FilterByCityObjIndex {
+    public:
+        /// MeshのうちCityObjectIndexが引数idに一致する箇所のみを取り出したMeshを生成して返します。
+        plateau::polygonMesh::Mesh filter(const plateau::polygonMesh::Mesh& src, polygonMesh::CityObjectIndex filter_id, const int uv4_atomic_index);
+    };
+}
