+++ src/granularity_convert/CMakeLists.txt
@@ -5,4 +5,5 @@ target_sources(plateau PRIVATE
         "convert_from_atomic_to_primary.cpp"
         "convert_from_atomic_to_area.cpp"
         "merge_primary_node_and_children.cpp"
+        "filter_by_city_obj_index.cpp"
 )
